Show interface selector on home page
If you just go to the hub (like jupyter.utoronto.ca), it actually
redirects you to https://jupyter.utoronto.ca/, it actually
redirects you to https://jupyter.utoronto.ca/hub/login?next=%2Fhub%2F.
So the 'next' URL is in fact set, and so we never see the
interface selector!

We don't want the selector on all hubs actually, as some might not
have RStudio nor Classic Notebook. So it's now explicitly
enableable on a per cluster basis.

Ref https://2i2c.freshdesk.com/a/tickets/97